pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jp"
version = "0.1.0"
description = "JobPilot - Clean and Maintainable Career Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["tests*", "logs*"]

[tool.black]
# No line length limit - let developers decide
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from exceptions - can be addressed later
    "B007",  # loop control variable not used - often intentional
    "B017",  # pytest.raises(Exception) - sometimes needed for testing
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "E203",  # whitespace before ':' (conflicts with black)
    "E402",  # module level import not at top - often needed
    "UP",    # pyupgrade rules - modernize later
    "PTH",   # pathlib modernization - can be addressed later
    "ARG",   # unused arguments - often intentional in interfaces
    "SIM",   # simplify code - can conflict with readability
    "C4",    # comprehension rules - can conflict with readability
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

# No line length limit - consistency with Black
# line-length = unlimited
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"test_*.py" = ["E402", "F841", "ARG001", "ARG002"]
"**/tests/**/*.py" = ["E402", "F841", "ARG001", "ARG002"]

[tool.ruff.isort]
known-first-party = ["backend"]
force-single-line = false
force-sort-within-sections = false
single-line-exclusions = ["typing"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config --verbose"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
